Created a lib module for aoc.
This lib module contains two functions: aoc_main, which is used to run each day's challenges as a standalone script, and parse, which is the former arg_parser.parse method. As a result, all days so far have been updated to use lib for their __main__ calls.

Additionally, a couple bugs with 2023/03.py was fixed. These include an issue where numbers at the very end of a table wouldn't be registered.